added classnames + moved countdown to ui
