+++ src/components/ui/countdown.tsx
@@ -2,12 +2,18 @@
 import { useState, useEffect } from "react";
 import data from "@/data/config";
 
-interface props {
+type CountdownClassNames = {
+  unit: string;
+  digit: string;
+  background: string;
+};
+interface digitProps {
   value: number;
   unit: string;
+  classNames: CountdownClassNames;
 }
 
-const Digits = ({ value, unit }: props) => {
+const Digits = ({ value, unit, classNames }: digitProps) => {
   return (
     <div className="flex flex-col items-center gap-4 last:hidden sm:last:flex">
       <div className="m-3 mb-0 flex gap-1 lg:!gap-1">
@@ -17,19 +23,23 @@ const Digits = ({ value, unit }: props) => {
           .split("")
           .map((digit, index) => (
             <div
-              className="flex items-center justify-center rounded-lg bg-white bg-opacity-40 p-3 text-lg font-bold text-white lg:min-w-11 lg:p-3 lg:text-3xl"
+              className={`flex items-center justify-center rounded-lg ${classNames.background} bg-opacity-40 p-3 text-lg font-bold ${classNames.digit} lg:min-w-11 lg:p-3 lg:text-3xl`}
               key={index}
             >
               {digit}
             </div>
           ))}
       </div>
-      <div className="m-2 mt-0 text-xs text-white">{unit}</div>
+      <div className={`m-2 mt-0 text-xs ${classNames.unit}`}>{unit}</div>
     </div>
   );
 };
 
-const Clock = () => {
+interface clockProps {
+  classNames: CountdownClassNames;
+}
+
+const Clock = ({ classNames }: clockProps) => {
   const [countdown, setCountdown] = useState({
     days: 0,
     hours: 0,
@@ -59,7 +69,7 @@ const Clock = () => {
   return (
     <div className="inline-flex font-bold">
       {Object.entries(countdown).map(([unit, value], index) => (
-        <Digits key={index} unit={unit} value={value} />
+        <Digits key={index} unit={unit} value={value} classNames={classNames} />
       ))}
     </div>
   );
